Bluetooth: Mesh: do not assume 32-bit architecture
Detected a mismatch when building tests with MSan for
native_sim/native/64

+++ subsys/bluetooth/mesh/crypto_psa.c
@@ -61,7 +61,7 @@ int bt_mesh_crypto_init(void)
 int bt_mesh_encrypt(const struct bt_mesh_key *key, const uint8_t plaintext[16],
 		    uint8_t enc_data[16])
 {
-	uint32_t output_len;
+	size_t output_len;
 	psa_status_t status;
 	int err = 0;
 
@@ -81,7 +81,7 @@ int bt_mesh_ccm_encrypt(const struct bt_mesh_key *key, uint8_t nonce[13],
 			const uint8_t *plaintext, size_t len, const uint8_t *aad,
 			size_t aad_len, uint8_t *enc_data, size_t mic_size)
 {
-	uint32_t output_len;
+	size_t output_len;
 	psa_status_t status;
 	int err = 0;
 	psa_algorithm_t alg = PSA_ALG_AEAD_WITH_SHORTENED_TAG(PSA_ALG_CCM, mic_size);
@@ -104,7 +104,7 @@ int bt_mesh_ccm_decrypt(const struct bt_mesh_key *key, uint8_t nonce[13],
 			const uint8_t *enc_data, size_t len, const uint8_t *aad,
 			size_t aad_len, uint8_t *plaintext, size_t mic_size)
 {
-	uint32_t output_len;
+	size_t output_len;
 	psa_status_t status;
 	int err = 0;
 	psa_algorithm_t alg = PSA_ALG_AEAD_WITH_SHORTENED_TAG(PSA_ALG_CCM, mic_size);
